Code
I used Python to filter the unstructured data from the reviews, detailed segmentation was achieved. The script separated the 5-star reviews from the 4-star reviews based on the rating. Furthermore, the reviews were categorized by their content, identifying whether the focus was on the food, the owner, both, or neither. Additionally, sentiment analysis was applied to classify each review as either positive or negative. This provided valuable insights into customer satisfaction and identified areas for improvement. Python was chosen for this analysis due to its ability to handle unstructured data efficiently and its suitability for tailored data analysis, ensuring that the results were both accurate and actionable. This segmentation is highly valuable for targeted marketing efforts, as it allows Changemaker to concentrate on the aspects of the business that resonate most with customers—whether it’s the food offerings, Marco’s personal interaction with customers, or a combination of both.