Add files via upload
Using Python, the reviews were segmented by star rating, categorizing them into groups such as 5-star and 4-star reviews. Additionally, a sentiment analysis function was applied to classify each review as either positive or negative based on its rating. This approach allows for the separation of customer feedback based on both star ratings and sentiment, enabling a more detailed analysis of customer satisfaction. Python was chosen for this data analysis because it provides a powerful and flexible toolset, allowing the team to perform tailored segmentation and sentiment analysis efficiently, ensuring the most relevant insights for the Changemaker project.