Documentation
This is the final report for class INF20030-Cloud Approaches for Enterprise Systems, where a group of 4 students developed an implementation plan for a Smart Urban Management System in Metroville City. The report provides a comprehensive analysis of Metroville City's path to operational enhancement through the adoption of a Smart Urban Management System (SUMS). It begins with an overview of Metroville's foundational operational domains, strengths, challenges, and growth vision then details the city's strategic implementation approach. It also identifies and prioritizes key business challenges, with a focus on how a SUMS solution can effectively address these issues. Based on this assessment, we outlines essential business requirements linked to specific SUMS functionalities, clarifying the potential impact of SUMS on Metroville's operations. To ensure an effective vendor selection,we came up with five recomendations aligned with Metroville's operational and implementation needs. In support of a successful SUMS rollout, the report also explores change management strategies and proposes contingency plans for both pre- and post-go-live phases. Further, it assesses SUMS's prospective contributions to Metroville's key operational areas, supporting the city's broader objectives. Lastly, we conduct examination on the potential benefits of upgrading Metroville's SUMS to SAP S/4HANA, considering both its capabilities and limitations. Based on case study insights and supplementary research, the report recommends this upgrade as a valuable step for Metroville.