Add files via upload
The main idea behind this research is to evaluate the adoption of the Scrum agile methodology within a subproject at Gothaer Insurance Group, particularly focusing on its integration into a traditionally waterfall environment. The research aims to identify the benefits and challenges of using Scrum in this context, and to explore potential improvements and alternatives.